requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pandas
numpy
scikit-learn
joblib
supabase
